+++ src/calandria/__init__.py
@@ -1,3 +1,3 @@
 """Calandria package."""
 
-__version__ = "2.5.2"
+__version__ = "2.5.3"